// frontend/core/shared/enums/questionnary_enum.ts
export enum Campaign {
  Education_Landscape = "Education Landscape",
}
export enum SectionName {
  Higher_edu_programmes = "Higher education programmes",
  Educational_framework = "Market-scale educational framework",
  Training_courses = "Short BIM training courses",
  Collaboration = "Collaboration between academia, government and industry",
}